+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "greek-meander"
-version = "0.1.1"
+version = "0.1.2"
 edition = "2024"
 description = "A Rust crate for creating rectangle border design of Greek Key (Meander)."
 license = "MIT"
